Adds bottomsheet content
