+++ domain/src/main/java/me/androidbox/domain/coin_detail/models/CoinFullDetailModel.kt
@@ -11,5 +11,6 @@ data class CoinFullDetailModel(
     val priceAt: Int,
     val symbol: String,
     val uuid: String,
-    val websiteUrl: String
+    val websiteUrl: String,
+    val color: String
 )
