+++ presentation/src/main/java/me/androidbox/presentation/coin_list/CoinListState.kt
@@ -10,5 +10,6 @@ data class CoinListState(
     val description: String = "",
     val websiteUrl: String = "",
     val marketCap: String = "",
+    val color: String = "",
     val isLoading: Boolean = false
 )
